refactor(analytics): align workbook ownership
